pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["RrylandD <RrylandD@gmail.com>"]
description = ""
name = "sittedbot"
version = "0.1.0"
[tool.poetry.dependencies]
discord = "^1.0"
python = "^3.7"
python-dotenv = "^0.13.0"
Flask = "^1.1"
[tool.poetry.dev-dependencies]
